requirements.lock.txt
#
# This file is autogenerated by pip-compile with Python 3.12
# by the following command:
#
#    pip-compile --no-strip-extras --output-file=requirements.lock.txt requirements.txt
#
annotated-doc==0.0.4
    # via fastapi
annotated-types==0.7.0
    # via pydantic
anyio==4.13.0
    # via
    #   httpx
    #   starlette
    #   watchfiles
bidict==0.23.1
    # via python-socketio
blinker==1.9.0
    # via flask
brotli==1.2.0
    # via geventhttpclient
certifi==2026.2.25
    # via
    #   geventhttpclient
    #   httpcore
    #   httpx
    #   requests
cffi==2.0.0
    # via cryptography
charset-normalizer==3.4.6
    # via requests
click==8.3.1
    # via
    #   flask
    #   uvicorn
cloudpickle==3.1.2
    # via shap
configargparse==1.7.5
    # via locust
coverage[toml]==7.13.5
    # via pytest-cov
cryptography==46.0.6
    # via google-auth
fastapi==0.135.2
    # via -r requirements.txt
flask==3.1.3
    # via
    #   flask-cors
    #   flask-login
    #   locust
flask-cors==6.0.2
    # via locust
flask-login==0.6.3
    # via locust
gevent==25.9.1
    # via
    #   geventhttpclient
    #   locust
geventhttpclient==2.3.9
    # via locust
google-api-core==2.30.0
    # via
    #   google-cloud-core
    #   google-cloud-storage
google-auth==2.49.1
    # via
    #   google-api-core
    #   google-cloud-core
    #   google-cloud-storage
google-cloud-core==2.5.0
    # via google-cloud-storage
google-cloud-storage==3.10.1
    # via -r requirements.txt
google-crc32c==1.8.0
    # via
    #   google-cloud-storage
    #   google-resumable-media
google-resumable-media==2.8.0
    # via google-cloud-storage
googleapis-common-protos==1.73.1
    # via google-api-core
greenlet==3.3.2
    # via gevent
h11==0.16.0
    # via
    #   httpcore
    #   uvicorn
    #   wsproto
httpcore==1.0.9
    # via httpx
httptools==0.7.1
    # via uvicorn
httpx==0.28.1
    # via -r requirements.txt
idna==3.11
    # via
    #   anyio
    #   httpx
    #   requests
iniconfig==2.3.0
    # via pytest
itsdangerous==2.2.0
    # via flask
jinja2==3.1.6
    # via flask
joblib==1.5.3
    # via scikit-learn
llvmlite==0.46.0
    # via
    #   numba
    #   shap
locust==2.43.3
    # via -r requirements.txt
markupsafe==3.0.3
    # via
    #   flask
    #   jinja2
    #   werkzeug
msgpack==1.1.2
    # via locust
numba==0.64.0
    # via shap
numpy==2.4.4
    # via
    #   -r requirements.txt
    #   numba
    #   pandas
    #   scikit-learn
    #   scipy
    #   shap
    #   xgboost
nvidia-nccl-cu12==2.29.7
    # via xgboost
packaging==26.0
    # via
    #   pytest
    #   shap
pandas==3.0.1
    # via shap
pluggy==1.6.0
    # via
    #   pytest
    #   pytest-cov
prometheus-client==0.24.1
    # via -r requirements.txt
proto-plus==1.27.2
    # via google-api-core
protobuf==6.33.6
    # via
    #   google-api-core
    #   googleapis-common-protos
    #   proto-plus
psutil==7.2.2
    # via locust
pyasn1==0.6.3
    # via pyasn1-modules
pyasn1-modules==0.4.2
    # via google-auth
pycparser==3.0
    # via cffi
pydantic==2.12.5
    # via
    #   -r requirements.txt
    #   fastapi
    #   pydantic-settings
pydantic-core==2.41.5
    # via pydantic
pydantic-settings==2.13.1
    # via -r requirements.txt
pygments==2.20.0
    # via pytest
pytest==9.0.2
    # via
    #   -r requirements.txt
    #   locust
    #   pytest-cov
pytest-cov==7.1.0
    # via -r requirements.txt
python-dateutil==2.9.0.post0
    # via pandas
python-dotenv==1.2.2
    # via
    #   -r requirements.txt
    #   pydantic-settings
    #   uvicorn
python-engineio==4.13.1
    # via
    #   locust
    #   python-socketio
python-socketio[client]==5.16.1
    # via locust
pyyaml==6.0.3
    # via uvicorn
pyzmq==27.1.0
    # via locust
requests==2.33.0
    # via
    #   google-api-core
    #   google-cloud-storage
    #   locust
    #   python-socketio
scikit-learn==1.8.0
    # via
    #   -r requirements.txt
    #   shap
scipy==1.17.1
    # via
    #   scikit-learn
    #   shap
    #   xgboost
shap==0.51.0
    # via -r requirements.txt
simple-websocket==1.1.0
    # via python-engineio
six==1.17.0
    # via python-dateutil
slicer==0.0.8
    # via shap
starlette==1.0.0
    # via fastapi
threadpoolctl==3.6.0
    # via scikit-learn
tqdm==4.67.3
    # via shap
typing-extensions==4.15.0
    # via
    #   anyio
    #   fastapi
    #   pydantic
    #   pydantic-core
    #   shap
    #   starlette
    #   typing-inspection
typing-inspection==0.4.2
    # via
    #   fastapi
    #   pydantic
    #   pydantic-settings
urllib3==2.6.3
    # via
    #   geventhttpclient
    #   requests
uvicorn[standard]==0.42.0
    # via -r requirements.txt
uvloop==0.22.1
    # via uvicorn
watchfiles==1.1.1
    # via uvicorn
websocket-client==1.9.0
    # via python-socketio
websockets==16.0
    # via uvicorn
werkzeug==3.1.7
    # via
    #   flask
    #   flask-cors
    #   flask-login
    #   locust
wsproto==1.3.2
    # via simple-websocket
xgboost==3.2.0
    # via -r requirements.txt
zope-event==6.1
    # via gevent
zope-interface==8.2
    # via gevent
